Model/Schema/SQLModel rework

+++ bin/test_company.sh
@@ -16,7 +16,7 @@ curl -s $BASE_URL/company/xyz | jq .
 #curl -s $BASE_URL/company/1 -X PUT -d '{"id": "3", "name": "New Corp"}' -H "Content-Type: application/json" | jq .
 
 echo "Create new"
-curl -s $BASE_URL/company -d '{"name": "New Corp", "github": "gh"}' -H "Content-Type: application/json" | jq . | tee /tmp/test_new_company.json
+curl -s $BASE_URL/company -d '{"name": "News Corp", "github": "gh"}' -H "Content-Type: application/json" | jq . | tee /tmp/test_new_company.json
 NEW_ID=`jq -r .id /tmp/test_new_company.json`
 echo NEW_ID $NEW_ID
 
@@ -32,5 +32,8 @@ curl -s $BASE_URL/company/5804cabc-11f9-43b0-a2b2-d6a966bdcf33 -X PUT -d '{"id":
 echo "Update"
 curl -s $BASE_URL/company/$NEW_ID -X PUT -d '{"id": "'"$NEW_ID"'", "name": "NCorp"}' -H "Content-Type: application/json" | jq .
 
-echo "Delete new"
-curl -s -X DELETE $BASE_URL/company/$NEW_ID | jq .
+if [ "$1" != "keep" ];
+then
+    echo "Delete new"
+    curl -s -X DELETE $BASE_URL/company/$NEW_ID | jq .
+fi
